refactor: more specific naming for blob data
