+++ server/routes/system-status.js
@@ -1,13 +1,13 @@
 const { notifyDashboardUri } = require('../config')
 const { scopes } = require('../permissions')
-const getStatusTable = require('../lib/view/get-status-table')
+const getDataItemsTable = require('../lib/view/get-data-items-table')
 
 module.exports = [
   {
     method: 'GET',
     path: '/system-status',
     handler: async (request, h) => {
-      const table = await getStatusTable()
+      const table = await getDataItemsTable()
       return h.view('system-status', { notifyDashboardUri, table })
     },
     options: {
